11_app/fastapi-demo/pyproject.toml
[project]
name = "fastapi-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "psycopg[binary]>=3.2.7",
    "python-dotenv>=1.1.0",
]
